% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{gg_par}
\alias{gg_par}
\title{Interpreter for graphical parameters}
\usage{
gg_par(..., stroke = NULL, pointsize = NULL)
}
\arguments{
\item{...}{Named arguments passed on to `gpar()`.}

\item{stroke}{Linewidth for points. Populates the `lwd` grid parameter.}

\item{pointsize}{Size for points. Populates the `fontsize` grid parameter.}
}
\value{
An object of class 'gpar'.
}
\description{
This is a wrapper for [`grid::gpar()`] that applies ggplot2's interpretation
of graphical parameters.
}
\keyword{internal}
