% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtintshade-package.R
\docType{package}
\name{ggtintshade-package}
\alias{ggtintshade}
\alias{ggtintshade-package}
\title{ggtintshade: Tinting and shading aesthetics for ggplot2}
\description{
Adds a 'tintshade' aesthetic to 'ggplot2' that varies the lightness of a mapped 'colour' or 'fill' within each hue group, keeping the plotted layer and its legend in sync. Supports both nested designs (each item belongs to one hue) and crossed designs (each lightness level appears across hues), the latter shown with a neutral greyscale legend.
}
\details{
The \code{ggtintshade} package is designed to provide an additional aesthetic to
typical \code{ggplot2} plots that allows colors to be lightened/darkened by
mapping the saturation/value coordinates as well as the hue. This allows for
grouped, nested, and crossed designs where not all colors should be equally
distinct.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wkumler/ggtintshade}
  \item Report bugs at \url{https://github.com/wkumler/ggtintshade/issues}
}

}
\author{
\strong{Maintainer}: William Kumler \email{wkumler@uw.edu}

}
\keyword{internal}
